Numbers Caster API 완성
힘들었다

+++ 250911/HobisLibrary/HobisLibraryDN472/Numbers/Int16Caster.cs
@@ -105,7 +105,7 @@
         #region [03) Unsigned]
         public static short From(uint un)
         {
-            if (un > Max)
+            if (un > (uint)Max)
             {
                 return Max;
             }
